+++ src/events/attributes.rs
@@ -514,18 +514,15 @@ pub struct Attributes<'a> {
     bytes: &'a [u8],
     /// Iterator state, independent from the actual source of bytes
     state: IterState,
-    /// Encoding used for `bytes`
-    decoder: Decoder,
 }
 
 impl<'a> Attributes<'a> {
     /// Internal constructor, used by `BytesStart`. Supplies data in reader's encoding
     #[inline]
-    pub(crate) const fn wrap(buf: &'a [u8], pos: usize, html: bool, decoder: Decoder) -> Self {
+    pub(crate) const fn wrap(buf: &'a [u8], pos: usize, html: bool) -> Self {
         Self {
             bytes: buf,
             state: IterState::new(pos, html),
-            decoder,
         }
     }
 
@@ -553,7 +550,7 @@ impl<'a> Attributes<'a> {
     /// assert_eq!(iter.next(), None);
     /// ```
     pub const fn new(buf: &'a str, pos: usize) -> Self {
-        Self::wrap(buf.as_bytes(), pos, false, Decoder::utf8())
+        Self::wrap(buf.as_bytes(), pos, false)
     }
 
     /// Creates a new attribute iterator from a buffer, allowing HTML attribute syntax.
@@ -578,7 +575,7 @@ impl<'a> Attributes<'a> {
     /// assert_eq!(iter.next(), None);
     /// ```
     pub const fn html(buf: &'a str, pos: usize) -> Self {
-        Self::wrap(buf.as_bytes(), pos, true, Decoder::utf8())
+        Self::wrap(buf.as_bytes(), pos, true)
     }
 
     /// Changes whether attributes should be checked for uniqueness.
@@ -666,22 +663,6 @@ impl<'a> Attributes<'a> {
             }
         })
     }
-
-    /// Get the decoder, used to decode bytes, read by the reader which produces
-    /// this iterator, to the strings.
-    ///
-    /// When iterator was created manually or get from a manually created [`BytesStart`],
-    /// encoding is UTF-8.
-    ///
-    /// If [`encoding`] feature is enabled and no encoding is specified in declaration,
-    /// defaults to UTF-8.
-    ///
-    /// [`BytesStart`]: crate::events::BytesStart
-    /// [`encoding`]: ../index.html#encoding
-    #[inline]
-    pub const fn decoder(&self) -> Decoder {
-        self.decoder
-    }
 }
 
 impl<'a> Debug for Attributes<'a> {
@@ -689,7 +670,6 @@ impl<'a> Debug for Attributes<'a> {
         f.debug_struct("Attributes")
             .field("bytes", &Bytes(self.bytes))
             .field("state", &self.state)
-            .field("decoder", &self.decoder)
             .finish()
     }
 }